requirements.txt
aiogram==3.26.0
anthropic>=0.84.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
PyMuPDF>=1.24.0
aiosqlite>=0.20.0
aiohttp>=3.9.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
loguru>=0.7.0
python-dateutil>=2.9.0
